+++ HISTORY.md
@@ -2,7 +2,8 @@ History of features added to LiChess Tools in time (reversed order)
 
 29 Apr
 
-- more piece sets and cleaning the old ones
+- more boards and piece sets and cleaning the old ones
+- Super Mario World sound theme
 
 28 Apr